Fix livetest error and compatibility with EventHub Track1 (#78)
* Update prefetch count in test

* Fix getting loop bug

* Lower logging level as too many logs will cause memory error in test pipeline

* Update prefetch count for the live test

* Update pretech count

* Update docstring of timeout type from int to float as part of the guideline

+++ uamqp/authentication/cbs_auth_async.py
@@ -177,7 +177,7 @@ class SASTokenAsync(SASTokenAuth, CBSAsyncAuthMixin):
     :type port: int
     :param timeout: The timeout in seconds in which to negotiate the token.
      The default value is 10 seconds.
-    :type timeout: int
+    :type timeout: float
     :param retry_policy: The retry policy for the PUT token request. The default
      retry policy has 3 retries.
     :type retry_policy: ~uamqp.authentication.cbs_auth.TokenRetryPolicy
@@ -224,7 +224,7 @@ class JWTTokenAsync(JWTTokenAuth, CBSAsyncAuthMixin):
     :type port: int
     :param timeout: The timeout in seconds in which to negotiate the token.
      The default value is 10 seconds.
-    :type timeout: int
+    :type timeout: float
     :param retry_policy: The retry policy for the PUT token request. The default
      retry policy has 3 retries.
     :type retry_policy: ~uamqp.authentication.cbs_auth.TokenRetryPolicy
